Fixing pulling from database issue with user_account phone/long and fixing...

See merge request campuses/cst-dal/may-2022/java/student-pairs/java-finalcapstone-team3!48

+++ java/src/main/java/com/techelevator/dao/JdbcUserAccountDao.java
@@ -88,7 +88,7 @@ public class JdbcUserAccountDao implements UserAccountDao {
         userAccount.setFirstName(rowSet.getString("f_name"));
         userAccount.setLastName(rowSet.getString("l_name"));
         userAccount.setEmail(rowSet.getString("email"));
-        userAccount.setPhone(BigInteger.valueOf(rowSet.getInt("phone")));
+        userAccount.setPhone(BigInteger.valueOf(rowSet.getLong("phone")));
         userAccount.setFlagged(rowSet.getBoolean("flagged"));
         return userAccount;
     }
